# openapi.yml
openapi: 3.0.0
info:
  title: FastCA
  description: |-
    FastCA REST API
  contact:
    email: ryan.bell62101@gmail.com
  license:
    name: MIT
    url: https://opensource.org/license/mit
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1
tags:
  - name: Authentication
    description: Authenticate using credentials or API keys
  - name: Certificate
    description: Certificate creation, deletion, and management
  - name: Certificate Authority
    description: Certificate authority creation, deletion, and management
  - name: Documentation
    description: OpenAPI documentation
paths:
  /openapi.yml:
    get:
      summary: OpenAPI YAML spec file
      tags:
        - Documentation
      responses:
        200:
          description: OpenAPI YAML spec file
          content: 
            application/yaml:
              schema:
              example: | 
                openapi: 3.0.0
                info:
                  title: FastCA
                  description: |-
                    FastCA REST API
  /docs:
    get:
      summary: SwaggerUI OpenAPI spec viewer
      tags:
        - Documentation
      responses:
        200:
          description: HTML Redoc page
          content:
            text/html:
              schema:

  /signin:
    post:
      summary: Sign in using username and password credentials
      tags:
        - Authentication
      requestBody:
        content:
          application/json: 
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        200:
          description: Successful signin
          headers:
            Set-Cookie:
              description: Sets the session cookie for an authenticated user
              schema:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
  /signout:
    post:
      summary: Sign out a currently authenticated user
      tags:
        - Authentication
      responses:
        200:
          description: Successful signin
          headers:
            Set-Cookie:
              description: Sets the session cookie for an authenticated user
              schema:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
      security: 
        - SessionAuth: []

  /ca/{caId}/fullchain:
    get:
      summary: Get the fullchain certificate for the certificate authority
      tags:
        - Certificate Authority
      parameters:
        - name: caId
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        200:
          description: Certificate authority fullchain certificate
          content:
            application/octet-stream: 
              schema:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
  /ca/list:
    get:
      summary: Get a list of certificate authorities managed by FastCA
      tags:
        - Certificate Authority
      responses:
        200:
          $ref: '#/components/responses/CertificateListResponse'
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
      parameters: 
        - name: limit
          schema:
            type: integer
            minimum: 5
            maximum: 100
            default: 20
          in: query
          required: false
        - name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          in: query
          required: false
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
  /ca:
    post:
      summary: Create a certificate authority
      tags:
        - Certificate Authority
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
      security:
        - ApiKeyAuth: []
        - SessionAuth: []

  /certificate/list:
    get:
      summary: Get a list of certificates managed by FastCA
      tags:
        - Certificate
      responses:
        200:
          $ref: '#/components/responses/CertificateListResponse'
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
      parameters: 
        - name: limit
          schema:
            type: integer
            minimum: 5
            maximum: 100
            default: 20
          in: query
          required: false
        - name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          in: query
          required: false
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
  /certificate:
    post:
      summary: Create a new certificate
      tags:
        - Certificate
      responses:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
  /certificate/{certId}:
    get:
      summary: Download a certificate
      tags:
        - Certificate
      responses:
        200:
          description: PEM encoded certificate files
          content:
            application/gzip:
              schema:
        4XX:
          description: Client error
          content:
            application/json:
              schema:
                type: string
      parameters: 
        - name: certId
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      security:
        - ApiKeyAuth: []
        - SessionAuth: []
components:
  responses:
    CertificateListResponse:
      description: List data for managed certificates
      content:
        application/json: 
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CertificatePaginatedData'
    CaListResponse:
      description: List of certificate authority datas
      content:
        application/json: 
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CaPaginatedData'
  schemas:
    CertificatePaginatedData:
      type: object
      required:
        - name
        - caName
        - certType
        - id
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 256
          example: Web Server Certificate
        caName:
          type: string
          minLength: 3
          maxLength: 256
          example: Web Server Certificate
        certType:
          type: string
          minLength: 3
          maxLength: 256
          example: Web Server Certificate
        id:
          type: integer
    CertificateData:
      type: object
      required:
        - name
        - subject
        - expires
        - ca
      properties:
        subject:
          type: object
          required:
            - organization
            - country
          properties:
            organization:
              type: string
            country:
              type: string
            province:
              type: string
            locality:
              type: string
            postalCode:
              type: string
            id:
              type: integer
        expires:
          type: string
          format: date-time
        ca:
          type: integer
          description: ID of the CA that signs this certificate
    CaPaginatedData:
      type: object
      required:
        - name
        - id
        - certType
      properties:
        name:
          minLength: 3
          maxLength: 256
          example: Web Server Certificate
        certType:
          type: string
          minLength: 3
          maxLength: 256
        id:
          type: integer
    CaData:
      type: object
      required:
        - name
        - subject
        - expires
      properties:
        subject:
          type: object
          required:
            - organization
            - country
          properties:
            organization:
              type: string
            country:
              type: string
            province:
              type: string
            locality:
              type: string
            postalCode:
              type: string
        expires:
          type: string
          format: date-time
    UserCredentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 5
          maxLength: 256
        password:
          type: string
          format: password
          minLength: 10
    UserSignupCredentials:
      type: object
      required:
        - username
        - email
        - password
      properties:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionAuth:
      type: apiKey
      in: cookie
      name: FASTCA_SESSION